MVKSync: Miscellaneous fixes.
Fix reversed condition in `MVKTimelineSemaphoreEmulated`, which probably
caused some test failures.

Don't increment the `MTLEvent` binary semaphore's counter unless a
command buffer is actually present to schedule a wait on.

Defer signal operations when a swapchain image is not yet available.
That way, the correct value will be signaled when the image is ready,
instead of causing GPU lockups and timeouts.

When a drawable is presented, immediately mark it available, instead of
waiting until the command buffer finishes. Otherwise, the wrong
semaphore could be signaled when an image is used twice in a row.

This should fix the problems using `MTLEvent` binary semaphores with
presentation, which was preventing us from enabling them by default when
available.

Special thanks to @apayen, whose
[idea](https://github.com/KhronosGroup/MoltenVK/issues/803) and
[change](https://github.com/apayen/MoltenVK/commit/a4ac715975ea3aa4cb95e91973b8f9010516304f)
were the basis for this.