MVKCmdWaitEvents: end current encoder before encodeWait
The Metal documentation for encodeWaitForEvent method says:

A command buffer can instruct the GPU to wait for an event only between passes,
not within a pass. If a command buffer has an active encoder, finish using the
encoder, call its endEncoding() method, and then call this method before
creating another encoder.

We should obey this rule and call endEncoding when there can be active encoder
around.