MVKCommandEncoderState: Mark vertex buffer binding as dirty when binding implicit buffer
In vertex shader, vertex attribute buffer and implicit buffer occupy the same
buffer binding space, but they are bound using different methods. When an
implicit buffer is used, it should mark the vertex (attribute) buffer at
the same buffer index as dirty so that it will be correctly bound when
activated.