MVKDescriptor: Prevent batched write combined image-sampler from overwriting each other in argument buffer
Currently samplerXXArray is translated to one textureXX_array and a single
sampler in MSL, consuming 2 bindings in Metal regardless of the array size.
This is already not an accurate translation. In the meantime, the descriptor
set update does not consider the difference in numbers between the approaches.
Despite the inaccuracy, the current solution "works" in some cases when not using
argument buffers. They are broken when argument buffer is enabled.

The underlying reason is that image (texture) and sampler are in the same index
space when used in argument buffers, and when updated in a batch (as sampler array)
using combined image-sampler descriptor type, the sampler is overwritten by the next
texture. This change avoids this kind of overwriting.

The ultimate solution would be properly implement texture/sampler arrays and account
for the semantic differences between Vulkan and Metal in descriptor sets. But let's
make it less broken first as the appropriate solution would require considerable effort.