MVKShaderLibrary: Handle specializtion with macros
The converted MSL may use macro instead of function constants to realize
spirv specialization constant for various reasons (e.g. when the constant
is used as array size).

In this case, we should define the macros at shader compilation stage and
generate different variants of the metal shader library depending on
macro-value mapping to make specializtion work properly when we cannot rely
on metal's specialization.

The current version is hacky as it scans the shader for macro each time we
create the function, and guesses the macro based on spirv-cross's habit.
Ideally, we should obtain the information from spirv-cross instead.