Store the residency set in MVKPhysicalDevice.
Since the MTLQueues are shared between MVKDevices, an application that creates
many devices (CTS, in particular) may run out of the number of residency sets
that can be added to a single MTLQueue.

This patch keeps a single residency set in MVKPhysicalDevice, which is added
once to each queue and shared as the queues are.

We could instead attach each MVKDevice's residency set to every created
MTLCommandBuffer, but that may incur a performance overhead of processing the
entries each time, which would affect all applications.

Fixes #2444